+++ lib/accept_headers/media_type.rb
@@ -8,6 +8,7 @@ module AcceptHeaders
     attr_reader :type, :subtype, :extensions
 
     MEDIA_TYPE_PATTERN = /^\s*(?<type>[\w!#$%^&*\-\+{}\\|'.`~]+)(?:\s*\/\s*(?<subtype>[\w!#$%^&*\-\+{}\\|'.`~]+))?\s*$/
+    PARAMS_PATTERN = /(?<attribute>[\w!#$%^&*\-\+{}\\|'.`~]+)\s*\=\s*(?:\"(?<value>[^"]*)\"|\'(?<value>[^']*)\'|(?<value>[\w!#$%^&*\-\+{}\\|\'.`~]*))/
 
     def initialize(type = '*', subtype = '*', q: 1.0, extensions: {})
       self.type = type
@@ -89,6 +90,37 @@ module AcceptHeaders
       else
         false
       end
+    end
+
+    def self.parse(header)
+      return nil if header.nil?
+      header.strip!
+      accept_media_range, accept_extensions = header.split(';', 2)
+      raise Error if accept_media_range.nil?
+      media_range = MEDIA_TYPE_PATTERN.match(accept_media_range)
+      raise Error if media_range.nil?
+      MediaType.new(
+        media_range[:type],
+        media_range[:subtype],
+        q: parse_q(accept_extensions),
+        extensions: parse_extensions(accept_extensions)
+      )
+    end
+
+    private
+    def self.parse_extensions(extensions_string)
+      return {} if !extensions_string || extensions_string.empty?
+      if extensions_string.match(/['"]/)
+        extensions = extensions_string.scan(PARAMS_PATTERN).map(&:compact).to_h
+      else
+        extensions = {}
+        extensions_string.split(';').each do |part|
+          param = PARAMS_PATTERN.match(part)
+          extensions[param[:attribute]] = param[:value] if param
+        end
+      end
+      extensions.delete('q')
+      extensions
     end
   end
 end